DocumentFaqs/requirements.txt
django = "*"
django-ninja = "*"
langchain = "*"
google-auth = "*"
aiohttp = "*"
langchain-core = "*"
langchain-text-splitters = "*"
langsmith = "*"
numpy = "*"
pydantic = "*"
pyyaml = "*"
requests = "*"
sqlalchemy = "*"
tenacity = "*"
openai = "*"
python-magic = "*"
python-docx = "*"
pymupdf = "*"
langchain-community = "*"
pypdf2 = "*"
langchain-google-genai = "*"
google-cloud-aiplatform = "*"
tiktoken = "*"